feat: Add updated_at field to User model
The User model in the Prisma schema has been updated to include an updated_at field. This field will be automatically set to the current date and time whenever a user record is created or updated. This change is necessary to track the last time a user record was modified.