chore: Update npm dependencies and add @types/qrcode
This commit updates the npm dependencies in the backend package.json file to include the latest stable versions of "@prisma/client" and "prisma". Additionally, it adds the "@types/qrcode" dependency. These updates ensure that the project is using the latest versions of the dependencies and includes the necessary type definitions for working with QR codes.