chore: Update .gitignore to ignore custom image files
This commit updates the .gitignore file in the backend directory to ignore custom image files located in the /images directory. By ignoring these files, they will not be tracked by Git and will not be included in the repository. This change helps to keep the repository clean and avoids unnecessary version control of custom image files.