chore: Update npm dependencies and add @zxcvbn-ts/core
This commit updates the npm dependencies in the backend and frontend package.json files to include the latest stable version of "@zxcvbn-ts/core". This dependency is added to enhance password strength checking and improve the security of the application. Updating the dependencies ensures that the project is using the latest stable version, which can include bug fixes, performance improvements, and new features.