chore: Update npm dependencies and add @elysiajs/cors
This commit updates the npm dependencies in the backend package.json file to include the latest stable version of "@elysiajs/cors". This dependency is added to enable Cross-Origin Resource Sharing (CORS) support in the application. Updating the dependencies ensures that the project is using the latest stable version, which can include bug fixes, performance improvements, and new features.